Add study preference to firefox.js.
This will be the preference that my code reads and listens for when it is turned on by Normandy to start up the study code.

While this is not strictly necessary, since 'Services.prefs.getBoolPref' allows you to set a default value if the pref is not present, this makes it easier for local testing/QA, since the pref will now show up in 'about:config' without Normandy.