Add TelemetryEvents class to ShieldStudySavant.jsm
See Issues #16 and #17 for how events will be registered.

Per chutten,
All event telemetry is disabled by default and the list of enabled categories is not persisted between sessions. This means any category to be collected must be enabled every time Firefox starts up. If a telemetry category is disabled (i.e. it is not explicitly enabled by my or anyone else's code), and I perform an action that would fire that event, it will NOT show up in about:telemetry.