Add study duration and expiration with testing override pref.
Testing override pref is an integer pref: "shield.savant.duration_override"; set values for this pref in ms.

Study duration is default set to 4 weeks per the PHD.

I have updated the TESTPLAN issue with instructions for testing expiration.

TODO: Resolve or use workaround
Expiration works as expected, however for the case when the study expired in between sessions (this is checked on SavantShieldStudy.init()), for some reason the 'end_study' event ping does not show up in 'about:telemetry'. It does for the case that the study expired during the session (this is checked on SavantShieldStudy.uninit()). I have asked in #telemetry IRC. A worst case scenario workaround if this is a bug is to only check for expiration on shutdown. My hypothesis is that recording an event for an enabled category and then disabling that category in the same session prevents any events for that category from being appended to the main ping...