Bug 1465703 - Add bookmark and follow_bookmark probes for Savant Shield study; r=adw, a=RyanVM
NOTE: THIS PATCH DOES NOT WORK FOR BOOKMARK ADD/REMOVE UNTIL A FIX IS APPLIED IN THE TAB PROBE PATCH.

These probes will register and record (for the duration of the study only):
* When a bookmark is added.
* When a bookmark is removed.
* When a bookmark is visited.

Also moved init of study module to later in startup as an idle task to mitigate performance impact. Previously, this patch was failing the browser/base/content/test/performance/browser_startup.js test.

Note for bookmark added/removed: Using the source argument, we can filter out bulk events like from Sync, import, restore, etc.

Note for bookmark visited: This will also fire for the case when the user visits a page directly that happens to be bookmarked. The user doesn't have to "click" on a bookmark.

MozReview-Commit-ID: EYVlTLfVLkd